chore(trading): market setup fixtures
