+++ tests/order_match/test_order_match.py
@@ -1,114 +1,9 @@
-import logging
 import pytest
 import re
-from collections import namedtuple
-from playwright.sync_api import expect
-
-# Defined namedtuples
-WalletConfig = namedtuple("WalletConfig", ["name", "passphrase"])
-
-# Wallet Configurations
-MM_WALLET = WalletConfig("mm", "pin")
-MM_WALLET2 = WalletConfig("mm2", "pin2")
-TERMINATE_WALLET = WalletConfig("FJMKnwfZdd48C8NqvYrG", "bY3DxwtsCstMIIZdNpKs")
-
-
-wallets = [MM_WALLET, MM_WALLET2, TERMINATE_WALLET]
-
-
-def setup_continuous_market(vega, page):
-    market_name = "BTC:DAI_Mar22"
-    logging.basicConfig(level=logging.INFO)
-
-    for wallet in wallets:
-        vega.create_key(wallet.name)
-
-    vega.mint(
-        MM_WALLET.name,
-        asset="VOTE",
-        amount=1e4,
-    )
-
-    vega.update_network_parameter(
-        MM_WALLET.name, parameter="market.fee.factors.makerFee", new_value="0.1"
-    )
-
-    vega.forward("10s")
-    vega.wait_for_total_catchup()
-
-    vega.create_asset(MM_WALLET.name, name="tDAI",
-                      symbol="tDAI", decimals=5, max_faucet_amount=1e10)
-    vega.wait_for_total_catchup()
-
-    tdai_id = vega.find_asset_id(symbol="tDAI")
-
-    vega.mint(
-        "Key 1",
-        asset=tdai_id,
-        amount=100e5,
-    )
-
-    vega.mint(
-        MM_WALLET.name,
-        asset=tdai_id,
-        amount=100e5,
-    )
-    vega.mint(
-        MM_WALLET2.name,
-        asset=tdai_id,
-        amount=100e5,
-    )
-
-    vega.wait_fn(10)
-    vega.wait_for_total_catchup()
-
-    vega.create_simple_market(
-        market_name,
-        proposal_key=MM_WALLET.name,
-        settlement_asset_id=tdai_id,
-        termination_key=TERMINATE_WALLET.name,
-        market_decimals=5,
-    )
-    vega.wait_for_total_catchup()
-
-    market_id = vega.all_markets()[0].id
-
-    vega.submit_simple_liquidity(
-        key_name=MM_WALLET.name,
-        market_id=market_id,
-        commitment_amount=10000,
-        fee=0.000,
-        reference_buy="PEGGED_REFERENCE_MID",
-        reference_sell="PEGGED_REFERENCE_MID",
-        delta_buy=1,
-        delta_sell=1,
-        is_amendment=False,
-    )
-
-    submit_order(vega, MM_WALLET.name, market_id, "SIDE_SELL", 1, 110)
-    submit_order(vega, MM_WALLET2.name, market_id, "SIDE_BUY", 1, 90)
-    submit_order(vega, MM_WALLET.name, market_id, "SIDE_SELL", 1, 105)
-    submit_order(vega, MM_WALLET2.name, market_id, "SIDE_BUY", 1, 95)
-
-    vega.wait_for_total_catchup()
-    vega.forward("10s")
-
-    page.goto(f"http://localhost:{vega.console_port}/#/markets/{market_id}")
-
-    submit_order(vega, "Key 1", market_id, "SIDE_BUY", 1, 110)
-
-
-def submit_order(vega, wallet_name, market_id, side, volume, price):
-    vega.submit_order(
-        trading_key=wallet_name,
-        market_id=market_id,
-        time_in_force="TIME_IN_FORCE_GTC",
-        order_type="TYPE_LIMIT",
-        side=side,
-        volume=volume,
-        price=price,
-
-    )
+from playwright.sync_api import expect, Page
+from vega_sim.service import VegaService
+from marketFixtures.openingAuctionMarket.openingAuctionMarket import setup_opening_auction_market
+from marketFixtures.continuousMarket.continuousMarket import setup_continuous_market
 
 
 # Could be turned into a helper function in the future.
@@ -168,9 +63,8 @@ def wait_for_graphql_response(page, query_name, timeout=5000):
 
 
 @pytest.mark.usefixtures("auth")
-def test_limit_order_trade_open_order(vega, page):
-    # setup continuous trading market with one user buy trade
-    setup_continuous_market(vega, page)
+def test_limit_order_trade_open_order(setup_opening_auction_market, page: Page):
+    setup_opening_auction_market
     # Assert that the user order is displayed on the orderbook
     orderbook_trade = page.get_by_test_id('price-11000000').nth(1)
 
@@ -194,13 +88,8 @@ def test_limit_order_trade_open_order(vega, page):
 
 
 @pytest.mark.usefixtures("auth")
-def test_limit_order_trade_open_position(vega, page):
-    # setup continuous trading market with one user buy trade
-    setup_continuous_market(vega, page)
-
-    vega.wait_for_total_catchup()
-    vega.forward("10s")
-
+def test_limit_order_trade_open_position(setup_continuous_market, page):
+    setup_continuous_market
     print("Assert Position:")
     # Assert that Position exists - Will fail if the order is incorrect.
     expected_position = [
@@ -222,12 +111,8 @@ def test_limit_order_trade_open_position(vega, page):
 
 
 @pytest.mark.usefixtures("auth")
-def test_limit_order_trade_order_trade_away(vega, page):
-    # setup continuous trading market with one user buy trade
-    setup_continuous_market(vega, page)
-
-    vega.wait_for_total_catchup()
-    vega.forward("10s")
+def test_limit_order_trade_order_trade_away(setup_continuous_market, page: Page):
+    setup_continuous_market
     # Assert that the order is no longer on the orderbook
     page.get_by_test_id('Orderbook').click()
     price_element = page.get_by_test_id('price-11000000').nth(1)
